requirements.txt
# ---- Foundation ------------------------------------------------------------
psycopg2-binary>=2.9.9         # PostgreSQL driver
python-dotenv>=1.0.1           # .env loading
requests>=2.31.0               # HTTP client for connectors

# ---- Google APIs (GSC, PageSpeed) ------------------------------------------
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ---- LLM -------------------------------------------------------------------
anthropic>=0.39.0              # Claude API client

# ---- Crawling (Technical SEO agent, Phase 1) -------------------------------
beautifulsoup4>=4.12.3
lxml>=5.1.0

# ---- Development -----------------------------------------------------------
# Kept here rather than a separate dev-requirements to keep setup one-step.
pytest>=8.0.0
